requirements.txt
streamlit
pandas
numpy
scikit-learn==1.6.1
joblib
pytest
pytest-cov
bandit
pip-audit

# API & backend
fastapi==0.110.0       
uvicorn[standard]==0.29.0
sqlalchemy==2.0.35
pydantic==1.10.15      
requests

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
prometheus-fastapi-instrumentator==5.11.1